backend/requirements.txt
aiohttp==3.11.11
channels==4.0.0
channels-redis==4.2.1
celery==5.3.5
cryptography>=45.0.5
Daphne==4.1.2
Django==4.2.18
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-shortuuidfield==0.1.3
djangorestframework==3.15.1
django-user-agents==0.4.0
django-simple-history==3.7.0
django-mptt==0.16.0
django_redis-6.0.0
exifread==3.0.0
httpx[http2]==0.25.2
imageio==2.35.1
python-dotenv==1.0.1
rawpy==0.22.0
eventlet==0.37.0
zipFly64>=1.3.1
pyyaml==6.0.1
ua-parser==0.18.0
user-agents==2.2.0
ujson==5.10.0
pydantic
requests